\name{Example.Data}
\alias{Example.Data}
\docType{data}
\title{
An example dataset
}
\description{
Example.Data is a hypothetical dataset constructed to demonstrate some of the functions in the package.}
\usage{data(Example.Data)}
\format{
A \code{data.frame} with \eqn{181} observations on \eqn{4} variables.
  \describe{
    \item{\code{Id}}{The Patient ID.}
    \item{\code{Treat}}{The treatment indicator, coded as \eqn{-1} = control and \eqn{1} = experimental.}
    \item{\code{T}}{The most credible outcome to assess therapeutic success.}
    \item{\code{S}}{The potential pretreatment predictor.}
  }
}
\keyword{Example.Data}
