\name{summary}
\alias{summary}
\alias{summary.GoodPretreatContCont}
\alias{summary.PCA.ContCont}
\alias{summary.Multivar.PCA.ContCont}
\alias{summary.Predict.Treat.ContCont}
\alias{summary.Predict.Treat.T0T1.ContCont}
\alias{summary.Predict.Treat.Multivar.ContCont}
\title{Summary}
\description{summary}
\usage{
\method{summary}{GoodPretreatContCont}(object, ..., Object)
\method{summary}{PCA.ContCont}(object, ..., Object)
\method{summary}{Predict.Treat.ContCont}(object, ..., Object)
\method{summary}{Predict.Treat.T0T1.ContCont}(object, ..., Object)
\method{summary}{Multivar.PCA.ContCont}(object, ..., Object)
\method{summary}{Predict.Treat.Multivar.ContCont}(object, ..., Object)
}
\keyword{internal}
